+++ ServiceLib/Response.cs
@@ -4,7 +4,7 @@ namespace ServiceLib
 {
     public class Response
     {
-        public Response(bool isValid, Vector vector, float confidence = 1)
+        public Response(bool isValid, Vector vector, double confidence = 1)
         {
             IsValid = isValid;
             Vector = vector;
@@ -13,6 +13,6 @@ namespace ServiceLib
 
         public bool IsValid { get; protected set; }
         public Vector Vector { get; protected set; }
-        public float Confidence { get; protected set; }
+        public double Confidence { get; protected set; }
     }
 }
